# docs/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Focus API 文档",
    "version": "1.0.0",
    "description": "Focus项目API接口文档",
    "contact": {
      "name": "Focus开发团队"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API V1 - 开发环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "TranscriptionTask": {
        "type": "object",
        "required": [
          "type",
          "input"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "offline"
            ],
            "description": "转录任务类型，目前支持离线转录"
          },
          "input": {
            "type": "object",
            "properties": {
              "sourceLanguage": {
                "type": "string",
                "enum": [
                  "cn",
                  "en"
                ],
                "description": "音频源语言，cn为中文，en为英文"
              },
              "fileUrl": {
                "type": "string",
                "description": "音频文件URL"
              }
            }
          },
          "parameters": {
            "type": "object",
            "properties": {
              "transcription": {
                "type": "object",
                "properties": {
                  "diarizationEnabled": {
                    "type": "boolean",
                    "description": "是否启用说话人分离"
                  },
                  "diarization": {
                    "type": "object",
                    "properties": {
                      "speakerCount": {
                        "type": "integer",
                        "description": "说话人数量"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/tingwu/upload": {
      "post": {
        "summary": "上传音频文件",
        "description": "上传音频文件并返回文件的URL，用于后续转录",
        "tags": [
          "通义听悟"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "audio": {
                    "type": "string",
                    "format": "binary",
                    "description": "音频文件（MP3、WAV等格式）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "文件上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "文件上传成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "fileName": {
                          "type": "string"
                        },
                        "fileUrl": {
                          "type": "string"
                        },
                        "size": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求无效，可能是缺少文件"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/tingwu/tasks": {
      "post": {
        "summary": "创建转录任务",
        "description": "创建音频转录任务，支持离线转录",
        "tags": [
          "通义听悟"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscriptionTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "转录任务创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "创建转录任务成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "taskId": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "无效的请求参数"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/tingwu/tasks/{taskId}": {
      "get": {
        "summary": "获取任务状态和结果",
        "description": "根据任务ID获取转录任务的状态和结果",
        "tags": [
          "通义听悟"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "转录任务ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取任务信息",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "获取任务信息成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "taskId": {
                          "type": "string"
                        },
                        "taskStatus": {
                          "type": "string",
                          "enum": [
                            "ONGOING",
                            "COMPLETED",
                            "FAILED"
                          ]
                        },
                        "result": {
                          "type": "object",
                          "properties": {
                            "transcription": {
                              "type": "string",
                              "description": "转录结果URL"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求无效，可能是缺少taskId"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/tingwu/tasks/{taskId}/poll": {
      "get": {
        "summary": "轮询任务结果直到完成",
        "description": "持续轮询任务状态直到任务完成或失败",
        "tags": [
          "通义听悟"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "转录任务ID"
          },
          {
            "in": "query",
            "name": "maxAttempts",
            "schema": {
              "type": "integer",
              "default": 30
            },
            "description": "最大轮询次数"
          },
          {
            "in": "query",
            "name": "interval",
            "schema": {
              "type": "integer",
              "default": 3000
            },
            "description": "轮询间隔(毫秒)"
          }
        ],
        "responses": {
          "200": {
            "description": "转录任务完成",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "转录任务完成"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "taskId": {
                          "type": "string"
                        },
                        "taskStatus": {
                          "type": "string"
                        },
                        "transcription": {
                          "type": "string",
                          "description": "转录结果URL"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求无效，可能是缺少taskId"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/tingwu/fetch-json": {
      "post": {
        "summary": "获取JSON文件内容",
        "description": "从指定URL获取JSON文件内容",
        "tags": [
          "通义听悟"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "description": "JSON文件的URL"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功获取JSON文件内容",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "获取JSON文件内容成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "获取到的JSON数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求无效，可能是缺少URL"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/tingwu/vocabularies": {
      "post": {
        "summary": "创建热词表",
        "description": "创建用于语音识别的热词表",
        "tags": [
          "通义听悟"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "words"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "热词表名称"
                  },
                  "words": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "热词列表"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "热词表创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "热词表创建成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "创建结果"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求无效，可能是缺少必要参数"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
...
  }
}